refactor: isolate actor compatibility views
